injected scripts added using evaluate, close #408
