+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 * Browser.text will return newlines on </br>
 * DefaultTimeout option on create browser
+* Injection scripts are now added through evaluate, so are not checked by CSP
+* Static config file removed
 
 2.5.1 / 2019-07-27
 ==================
